# api/openapi.yaml
openapi: 3.0.3
info:
  title: happyDomain API
  description: Finally a simple interface for domain names.
  version: "0.1.0"
  contact:
    email: contact+api@happydomain.org
    name: happyDomain team
  license:
    name: GNU Affero General Public License v3.0 or later
    url: https://spdx.org/licenses/AGPL-3.0-or-later.html
...
servers:

  - url: http://localhost:8081/api
    description: Development server

  - url: https://app.happydomain.org/api
    description: Main production server
paths:
...
  /api/domains/{domainId}/zone/{zoneId}/{subdomain}:

    get:
      operationId: list_subdomain_services
      description: Returns the services associated with the given subdomain.
      parameters:
      - description: Domain identifier
        in: path
        name: domainId
        required: true
        schema:
          type: string
      - description: Zone identifier
        in: path
        name: zoneId
        required: true
        schema:
          type: string
      - description: Part of the subdomain considered for the service (@ for the root
          of the zone ; subdomain is relative to the root, do not include it)
        in: path
        name: subdomain
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: './schemas.yaml#/components/schemas/ZoneServices'
        "401":
          description: Authentication failure
          content:
            application/json:
              schema:
                $ref: './schemas.yaml#/components/schemas/ErrorResponse'
        "404":
          description: Domain or Zone not found
          content:
            application/json:
              schema:
                $ref: './schemas.yaml#/components/schemas/ErrorResponse'
      security:
        - ApiKeyAuth: []
      summary: List services
      tags:
        - zones
...
  /api/domains/{domainId}/zone/{zoneId}/apply_changes:

    post:
      operationId: apply_zone_changes
      description: Perform the requested changes with the provider.
      parameters:
      - description: Domain identifier
        in: path
        name: domainId
        required: true
        schema:
          type: string
      - description: Zone identifier
        in: path
        name: zoneId
        required: true
        schema:
          type: string
      requestBody:
        description: Differences (from /diff_zones) to apply
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                type: string
      responses:
        "200":
          description: The new Zone metadata containing the current zone
          content:
            application/json:
              schema:
                $ref: './schemas.yaml#/components/schemas/ZoneMeta'
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: './schemas.yaml#/components/schemas/ErrorResponse'
        "401":
          description: Authentication failure
          content:
            application/json:
              schema:
                $ref: './schemas.yaml#/components/schemas/ErrorResponse'
        "404":
          description: Domain or Zone not found
          content:
            application/json:
              schema:
                $ref: './schemas.yaml#/components/schemas/ErrorResponse'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: './schemas.yaml#/components/schemas/ErrorResponse'
      security:
        - ApiKeyAuth: []
      summary: Performs requested changes to the real zone.
      tags:
        - zones
  /api/domains/{domainId}/zone/{zoneId}/diff/{oldZoneId}:

    post:
      operationId: diff_zones
      description: Compute the difference between the two zone identifiers given.
      parameters:
      - description: Domain identifier
        in: path
        name: domainId
        required: true
        schema:
          type: string
      - description: Zone identifier to use as the new one.
        in: path
        name: zoneId
        required: true
        schema:
          type: string
      - description: Zone identifier to use as the old one. Currently only @ are expected,
          to use the currently deployed zone.
        in: path
        name: oldZoneId
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Differences, reported as text, one diff per item
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: './schemas.yaml#/components/schemas/Correction'
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: './schemas.yaml#/components/schemas/ErrorResponse'
        "401":
          description: Authentication failure
          content:
            application/json:
              schema:
                $ref: './schemas.yaml#/components/schemas/ErrorResponse'
        "404":
          description: Domain not found
          content:
            application/json:
              schema:
                $ref: './schemas.yaml#/components/schemas/ErrorResponse'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: './schemas.yaml#/components/schemas/ErrorResponse'
        "501":
          description: Diff between to zone identifier, currently not supported
          content:
            application/json:
              schema:
                $ref: './schemas.yaml#/components/schemas/ErrorResponse'
      security:
        - ApiKeyAuth: []
      summary: Compute differences between zones.
      tags:
        - zones
  /api/domains/{domainId}/zone/{zoneId}/records:

    patch:
      operationId: update_zone_record
      description: Update a given record in the zone.
      parameters:
      - description: Domain identifier
        in: path
        name: domainId
        required: true
        schema:
          type: string
      - description: Zone identifier
        in: path
        name: zoneId
        required: true
        schema:
          type: string
      requestBody:
        description: Record to update as text
        required: true
        content:
          application/json:
            schema:
              $ref: './schemas.yaml#/components/schemas/UpdateRecordForm'
      responses:
        "200":
          description: The updated zone
          content:
            application/json:
              schema:
                $ref: './schemas.yaml#/components/schemas/Zone'
        "401":
          description: Authentication failure
          content:
            application/json:
              schema:
                $ref: './schemas.yaml#/components/schemas/ErrorResponse'
        "404":
          description: Domain or Zone not found
          content:
            application/json:
              schema:
                $ref: './schemas.yaml#/components/schemas/ErrorResponse'
      security:
        - ApiKeyAuth: []
      summary: Update a given record in the zone.
      tags:
        - zones

    post:
      operationId: add_zone_records
      description: Add a given record in the zone.
      parameters:
      - description: Domain identifier
        in: path
        name: domainId
        required: true
        schema:
          type: string
      - description: Zone identifier
        in: path
        name: zoneId
        required: true
        schema:
          type: string
      requestBody:
        description: Records to add as text, one record per line array
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                type: string
      responses:
        "200":
          description: The updated zone
          content:
            application/json:
              schema:
                $ref: './schemas.yaml#/components/schemas/Zone'
        "401":
          description: Authentication failure
          content:
            application/json:
              schema:
                $ref: './schemas.yaml#/components/schemas/ErrorResponse'
        "404":
          description: Domain or Zone not found
          content:
            application/json:
              schema:
                $ref: './schemas.yaml#/components/schemas/ErrorResponse'
      security:
        - ApiKeyAuth: []
      summary: Add a given record in the zone.
      tags:
        - zones
  /api/domains/{domainId}/zone/{zoneId}/records/delete:

    post:
      operationId: delete_zone_records
      description: Delete a given record in the zone.
      parameters:
      - description: Domain identifier
        in: path
        name: domainId
        required: true
        schema:
          type: string
      - description: Zone identifier
        in: path
        name: zoneId
        required: true
        schema:
          type: string
      requestBody:
        description: Records to delete as text, one record per line array
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                type: string
      responses:
        "200":
          description: The updated zone
          content:
            application/json:
              schema:
                $ref: './schemas.yaml#/components/schemas/Zone'
        "401":
          description: Authentication failure
          content:
            application/json:
              schema:
                $ref: './schemas.yaml#/components/schemas/ErrorResponse'
        "404":
          description: Domain or Zone not found
          content:
            application/json:
              schema:
                $ref: './schemas.yaml#/components/schemas/ErrorResponse'
      security:
        - ApiKeyAuth: []
      summary: Delete a given record in the zone.
      tags:
        - zones
  /api/domains/{domainId}/zone/{zoneId}/view:

    post:
      operationId: view_zone
      description: Create a flatten export of the zone that can be read as a BIND-like
        file.
      parameters:
      - description: Domain identifier
        in: path
        name: domainId
        required: true
        schema:
          type: string
      - description: Zone identifier
        in: path
        name: zoneId
        required: true
        schema:
          type: string
      responses:
        "200":
          description: The exported zone file (with initial and leading JSON quote)
          content:
            application/json:
              schema:
                type: string
        "401":
          description: Authentication failure
          content:
            application/json:
              schema:
                $ref: './schemas.yaml#/components/schemas/ErrorResponse'
        "404":
          description: Domain or Zone not found
          content:
            application/json:
              schema:
                $ref: './schemas.yaml#/components/schemas/ErrorResponse'
      security:
        - ApiKeyAuth: []
      summary: Get flatten zone file.
      tags:
        - zones
...
components:
  securitySchemes:

    ApiKeyAuth:
      description: Description for what is this security definition being used
      in: header
      name: Authorization
      type: apiKey

    BasicAuth:
      type: http
      scheme: basic
...